user_side/requirements.txt
auto-sklearn==0.15.0
TPOT==0.11.7
FLAML==1.0.0
lightautoml==0.3.7.3
scikit-learn==0.24.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
pandas==1.4.3
numpy==1.21.6
xgboost==1.3.3
lightgbm==3.2.1
